+++ lib/QECCore/src/QECCore.jl
@@ -14,10 +14,13 @@ export AbstractECC, AbstractQECC, AbstractCECC, AbstractCSSCode, AbstractDistanc
 export Perfect5, Cleve8, Gottesman
 
 # CSS Codes
-export Toric, Bitflip3, Phaseflip3, Shor9, Steane7, Surface, CSS, QuantumReedMuller, Triangular488, Triangular666, DelfosseReichardtRepCode
+export Toric, Bitflip3, Phaseflip3, Shor9, Steane7, Surface, CSS, QuantumReedMuller, Triangular488, Triangular666, DelfosseReichardt,  DelfosseReichardtRepCode
 
 # Classical Codes
 export RepCode, ReedMuller, RecursiveReedMuller, Golay, Hamming
+
+# utilities
+export search_self_orthogonal_rm_codes
 
 include("interface.jl")
 include("codes/util.jl")
@@ -45,6 +48,7 @@ include("codes/classical/recursivereedmuller.jl")
 include("codes/quantum/quantumreedmuller.jl")
 
 # Delfosse-Reichardt Codes
+include("codes/quantum/delfosse_reichardt_code.jl")
 include("codes/quantum/delfosse_reichardt_repcode.jl")
 
 function __init__()