+++ lib/QECCore/test/test_codes.jl
@@ -62,4 +62,8 @@
     @testset "Delfosse-Reichardt Repetition Code" begin
         include("codes/delfosse_reichardt_repcode.jl")
     end
+
+    @testset "Delfosse-Reichardt Code" begin
+        include("codes/delfosse_reichardt_code.jl")
+    end
 end