+++ CHANGELOG.md
@@ -13,7 +13,9 @@
 - Drop support for Julia 1.9.
 - Add convenience wrappers for code families: `honeycomb_color_codes` and `haah_cubic_codes` via `LPCode` construction.
 - Some codes are moved to `QECCore` from `QuantumClifford`, including `Hamming`, `Golay`, `Triangular488 `, `Triangular666 `, `Gottesman`.
-- Add `[[4p, 2(p − 2), 4]]` Delfosse-Reichardt repetition code.
+- Add `Delfosse-Reichardt` codes from classical self-orthogonal `Reed-Muller` seed codes to `QECCore`.
+- Add `[[4p, 2(p − 2), 4]]` Delfosse-Reichardt repetition code to `QECCore`.
+
 
 ### Private API